+++ SharedUtils/src/test/java/com/braintreepayments/api/HttpRequestUnitTest.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Locale;
@@ -17,9 +18,9 @@ import static org.junit.Assert.*;
 
 
 @RunWith(Enclosed.class)
-public class HttpRequestTest {
+public class HttpRequestUnitTest {
 
-    public static class NonParameterizedHttpRequestTests {
+    public static class NonParameterizedScenarios {
 
         @Test
         public void getPath_returnsPath() {
@@ -34,7 +35,24 @@ public class HttpRequestTest {
             HttpRequest sut = HttpRequest.newInstance()
                     .data("sample data");
 
-            assertEquals("sample data", sut.getData());
+            assertEquals("sample data", new String(sut.getData(), StandardCharsets.UTF_8));
+        }
+
+        @Test
+        public void dispose_whenDataIsNull_doesNothing() {
+            HttpRequest sut = HttpRequest.newInstance();
+            sut.dispose();
+            assertNull(sut.getData());
+        }
+
+        @Test
+        public void dispose_whenDataExists_zeroesOutData() {
+            HttpRequest sut = HttpRequest.newInstance()
+                    .data("sample data");
+            sut.dispose();
+
+            byte[] actual = sut.getData();
+            assertArrayEquals(new byte[actual.length], actual);
         }
 
         @Test
@@ -115,13 +133,13 @@ public class HttpRequestTest {
     }
 
     @RunWith(Parameterized.class)
-    public static class GetURLTest {
+    public static class GetURLScenarios {
 
         private final String baseUrl;
         private final String path;
         private final URL expectedURL;
 
-        public GetURLTest(String baseUrl, String path, URL expectedURL) {
+        public GetURLScenarios(String baseUrl, String path, URL expectedURL) {
             this.baseUrl = baseUrl;
             this.path = path;
             this.expectedURL = expectedURL;