+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Braintree Android SDK Release Notes
 
 ## unreleased
+
 * GooglePay
   * Deprecate `googleMerchantId`
+* SharedUtils
+  * Use byte array to hold `HttpRequest` data. Dispose data immediately after making http request.
 
 ## 4.8.0